% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESTNUMB.R
\docType{data}
\name{TESTNUMB}
\alias{TESTNUMB}
\title{A data set created by merging four smaller data sets.  Three of those smaller data sets are data from
three surveys (O1, O2, O3); the other is data from a "gold standard" survey (A1).  All four smaller data
sets consist of the same three variables (Q1, Q2, Q3): responses to the same three questions, asked
by each survey from the same 10 respondents (ID), along the same 1-99 response scale.}
\format{A data frame with 10 rows and 13 variables
\describe{\item{ID, AQ1, AQ2, AQ3, O1Q1, O1Q2, O1Q3, O2Q1, O2Q2, O2Q3, O3Q1, O3Q2, O3Q3}{sets of three variables from each of four surveys, merged together}
}}
\source{
Example data generated by author
}
\usage{
TESTNUMB
}
\description{
A data set created by merging four smaller data sets.  Three of those smaller data sets are data from
three surveys (O1, O2, O3); the other is data from a "gold standard" survey (A1).  All four smaller data
sets consist of the same three variables (Q1, Q2, Q3): responses to the same three questions, asked
by each survey from the same 10 respondents (ID), along the same 1-99 response scale.
}
\keyword{datasets}
